simplify some engine headers
