+++ M2/Macaulay2/e/interface/computation.h
@@ -1,6 +1,8 @@
 #ifndef _computation_h_
 #  define _computation_h_
 
+#  include "engine-includes.hpp"
+
 #  if defined(__cplusplus)
 extern "C" {
 #  endif
@@ -28,6 +30,21 @@ enum ComputationStatusCode {
   COMP_DONE_SUBRING_LIMIT = 16, /* SubringLimit */
   COMP_COMPUTING = 17,          /* computing */
   COMP_OVERFLOWED = 18,         /* overflowed */
+};
+
+struct StopConditions
+{
+  M2_bool always_stop;
+  M2_bool stop_after_degree;
+  M2_arrayint degree_limit;  // Stop after completing this 'slanted' degree
+  unsigned int basis_element_limit;  // Number of gb elements
+  unsigned int syzygy_limit;
+  unsigned int pair_limit;
+  M2_bool use_codim_limit;
+  unsigned int codim_limit;
+  unsigned int subring_limit;
+  M2_bool just_min_gens;
+  M2_arrayint length_limit;  // ignored for GB computations
 };
 
 // TODO: should these move elsewhere?
