+++ M2/Macaulay2/e/interface/freemodule.cpp
@@ -3,7 +3,6 @@
 #include "interface/freemodule.h"
 
 #include "buffer.hpp"
-#include "engine-includes.hpp"
 #include "error.h"
 #include "exceptions.hpp"
 #include "freemod.hpp"
