+++ M2/Macaulay2/e/interface/groebner.h
@@ -33,7 +33,7 @@ typedef struct RingMap RingMap;
 typedef EngineComputation EngineComputationOrNull;
 
 /**
-   Groebner computations interface
+   Groebner computations interface routines
  */
 
 #  if defined(__cplusplus)
