+++ M2/Macaulay2/e/interface/ringmap.cpp
@@ -1,7 +1,5 @@
 #include "interface/ringmap.h"
 
-#include "engine-includes.hpp"
-
 #include "buffer.hpp"
 #include "error.h"
 #include "exceptions.hpp"
